Support non-absolute executable paths in BuildProcessChecker pattern
Right now the pattern assumed that executables were launched with an absolute
path (it required a slash or backslash to the left of the name). It's common
for tools like make or cmake to be called without a path though, so allow the
executable name to be anchored left instead.